move shims to new shims module
